Reworks how the filelists work with the project variables.
Now the Jagati is opionated based on the 6 project directory variables and will presume
file locations are relative to one of these directories based on which macro is used.
The dox macro will check if the file exists first so the file can be in the header
folder, and if it is it will be allowed.