Now supports file lists, automated opinionated linking, better error messages and more.
The test suite now has 400 assertions. Most testing features related to FileLists.
Filelists are how the Jagati keeps track of source files in each package and what
will be part of the build in each. When adding files to different lists there is
a collection of AddXxxxFile macros that each work on different lists. These macros
check source in a sane folder or produces errors.